feat: added input mock
