--- src/components/Form/Input/mock.ts
@@ -1,26 +1,0 @@
-import { InitialValueType, ReducerActionType } from '~/@types/Form';
-
-export function reducerMock(state: InitialValueType, action: ReducerActionType) {
-  const value = action.payload.value;
-  const error = action.payload.error;
-
-  switch (action.type) {
-    case 'RECIPE_NAME':
-      return { ...state, recipeName: { value, error } };
-    case 'USERNAME':
-      return { ...state, username: { value, error } };
-    case 'EMAIL':
-      return { ...state, email: { value, error } };
-    case 'PASSWORD':
-      return { ...state, password: { value, error } };
-    default:
-      return state;
-  }
-}
-
-export const initialValueMock: InitialValueType = {
-  recipeName: { value: '', error: null },
-  username: { value: '', error: null },
-  email: { value: '', error: null },
-  password: { value: '', error: null },
-};
